move EcrDisplay and RelatedEcr to separate types.ts to avoid circular imports
