+++ containers/ecr-viewer/src/app/services/listEcrDataService.ts
@@ -9,19 +9,13 @@ import {
 import { getDb } from "@/app/data/metadataDb/database";
 import { getSql } from "@/app/data/metadataDb/dialects/common";
 import { ecr_data, Core, User } from "@/app/data/metadataDb/types/core";
+import { EcrDisplay, RelatedEcr } from "@/app/types";
 import { DateRangePeriod } from "@/app/utils/date-utils";
 
 import { audit } from "./auditLogService";
 import { UserFacingError } from "./errorService";
 import { formatDate, formatDateTime } from "./formatDateService";
 import { getLoggedInUser } from "./loggedInUserService";
-
-export interface RelatedEcr {
-  eicr_id: string;
-  date_created: Date;
-  eicr_version_number: string | undefined;
-  set_id: string;
-}
 
 export interface MetadataModel {
   eicr_id: string;
@@ -35,19 +29,6 @@ export interface MetadataModel {
   last_name: string | undefined;
   birth_date: Date | undefined;
   encounter_start_date: Date | undefined;
-}
-export interface EcrDisplay {
-  ecrId: string;
-  patient_first_name: string;
-  patient_last_name: string;
-  patient_date_of_birth: string | undefined;
-  reportable_conditions: string[];
-  rule_summaries: string[];
-  patient_report_date: string;
-  date_created: string;
-  eicr_set_id: string | undefined;
-  eicr_version_number: string | undefined;
-  related_ecrs: RelatedEcr[];
 }
 
 /**
